fix: JWT permission issues (#4)
Description
---
* Propagate the user accepted permissions to the signaling server
* Adapt to the new permission types for transactions and account info

Motivation and Context
---
Up until now on web applications that use the `tari-connector` project
to interact with a wallet daemon, the tokens being generated did not
work. https://github.com/tari-project/tari-dan/pull/549 is the
corresponding PR in the `tari-dan` side.

How Has This Been Tested?
---
Manually using a web application that uses the `tari-connector` to
interact with the wallet daemon

What process can a PR reviewer use to test or verify this change?
---
Manually use a web application that uses the `tari-connector` to
interact with the wallet daemon