feat: new permission for public key listing (#7)
In the context of atomic swaps, and in general in any web that may
result in the user creating a new account, we want to be able to list
all the public keys of the wallet.

The public key list also indicates which key is the default key, so that
we can use it in the web interface to predict the component address of
the newly created user account.

This PR adds the representation for the new type of permission
(`TariPermissionKeyList`) needed for public key listing.

There is a [corresponding
PR](https://github.com/tari-project/tari-dan/pull/647) in the `tari-dan`
repository