feat: get and store wallet daemon's JWT after WebRTC channel is open (#9)
Up until now, `tari-connector` used the signaling server's JWT to call
the wallet daemon. This worked because the signaling server and wallet
daemon to use the same JWT secret.

This is will no longer be the case, so to solve it this PR fetches and
store the wallet daemon's JWT after the WebRTC channel is open (i.e.
after `setAnswer` is called). The web developer does not need to do or
change anything, existing web apps with `tari-connector` will continue
to work as the TariConnection.token is abstracted away.