feat: implement signaling server polling  (#10)
Up until now, users had to manually click (or the web developers had to
implement a polling) on a button to call the `setAnswer` method, that
tries to set up the webRTC peer connection.

This PR implements a polling on the `tari-connector` itself, so after
the user clicks on the connection button (and the QR is shown) every 2
seconds it tries to call `setAnswer` until the connection is
established.

Web developers need to implement a new `onConnection` callback that will
be called by the `tari-connector` once the WebRTC connection with the
wallet (and the wallet JWT) is ready.